Extracts show function via hook
